move cmd.getID -> util.MakeID
